requirements.txt
torch>=1.4.0
torchvision>=0.5.0
dominate>=2.4.0
visdom>=0.1.8.8
wandb
tqdm
einops
pytorch-msssim
numba
pandas
pillow<10.0
numpy
matplotlib
datasets
kornia
